Refactor file paths and imports
