+++ src/containers/cashier_dashboard_payment_confirm_popup/ConfirmPaymentPopUp.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import HorizontalDivider from '../../components/divider/HorizontalDivider';
+import HorizontalDivider from '../../shared/divider/HorizontalDivider';
 import MedicineGridPopUp from '../../features/cashier-dashboard/components/order-confirm/MedineGridPopUp';
 type Props = {
   isCardVisible: boolean;
