+++ src/features/cashier-dashboard/components/sidebar/CashierSideBar.tsx
@@ -5,7 +5,7 @@ import { PiPillLight } from 'react-icons/pi';
 import { IoIosFitness } from 'react-icons/io';
 import { PiFirstAidKit } from 'react-icons/pi';
 import { CiMedicalClipboard } from 'react-icons/ci';
-import ButtonWithIconAndTextVertical from '../../../../components/buttons/ButtonWithIconAndTextVertical';
+import ButtonWithIconAndTextVertical from '../../../../shared/buttons/ButtonWithIconAndTextVertical';
 
 type Props = {};
