+++ src/shared/navbar/CashierNavBar.tsx
